pyproject.toml
[project]
name = "stock-akshare"
version = "0.1.0"
description = ""
authors = [
    {name = "Shawn Sang",email = "shawn.sang@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "akshare (>=1.16.75,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-html = "^4.1.1"
pytest-cov = "^6.1.1"

